move testutil package to fix import cycle
