+++ internal/engine/podlogmanager_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	v1 "k8s.io/api/core/v1"
 
-	"github.com/windmilleng/tilt/internal/testutils"
+	"github.com/windmilleng/tilt/internal/testutils/manifestutils"
 
 	"github.com/windmilleng/tilt/internal/container"
 	"github.com/windmilleng/tilt/internal/k8s"
@@ -33,7 +33,7 @@ func TestLogs(t *testing.T) {
 
 	state := f.store.LockMutableStateForTesting()
 	state.WatchFiles = true
-	state.UpsertManifestTarget(testutils.NewManifestTargetWithPod(
+	state.UpsertManifestTarget(manifestutils.NewManifestTargetWithPod(
 		model.Manifest{Name: "server"},
 		store.Pod{
 			PodID:         podID,
@@ -55,7 +55,7 @@ func TestLogActions(t *testing.T) {
 
 	state := f.store.LockMutableStateForTesting()
 	state.WatchFiles = true
-	state.UpsertManifestTarget(testutils.NewManifestTargetWithPod(
+	state.UpsertManifestTarget(manifestutils.NewManifestTargetWithPod(
 		model.Manifest{Name: "server"},
 		store.Pod{
 			PodID:         podID,
@@ -77,7 +77,7 @@ func TestLogsFailed(t *testing.T) {
 
 	state := f.store.LockMutableStateForTesting()
 	state.WatchFiles = true
-	state.UpsertManifestTarget(testutils.NewManifestTargetWithPod(
+	state.UpsertManifestTarget(manifestutils.NewManifestTargetWithPod(
 		model.Manifest{Name: "server"},
 		store.Pod{
 			PodID:         podID,
@@ -100,7 +100,7 @@ func TestLogsCanceledUnexpectedly(t *testing.T) {
 
 	state := f.store.LockMutableStateForTesting()
 	state.WatchFiles = true
-	state.UpsertManifestTarget(testutils.NewManifestTargetWithPod(
+	state.UpsertManifestTarget(manifestutils.NewManifestTargetWithPod(
 		model.Manifest{Name: "server"},
 		store.Pod{
 			PodID:         podID,
@@ -129,7 +129,7 @@ func TestMultiContainerLogs(t *testing.T) {
 
 	state := f.store.LockMutableStateForTesting()
 	state.WatchFiles = true
-	state.UpsertManifestTarget(testutils.NewManifestTargetWithPod(
+	state.UpsertManifestTarget(manifestutils.NewManifestTargetWithPod(
 		model.Manifest{Name: "server"},
 		store.Pod{
 			PodID:         podID,
@@ -164,7 +164,7 @@ func TestContainerPrefixes(t *testing.T) {
 
 	state := f.store.LockMutableStateForTesting()
 	state.WatchFiles = true
-	state.UpsertManifestTarget(testutils.NewManifestTargetWithPod(
+	state.UpsertManifestTarget(manifestutils.NewManifestTargetWithPod(
 		model.Manifest{Name: "multiContainer"},
 		// Pod with multiple containers -- logs should be prefixed with container name
 		store.Pod{
@@ -177,7 +177,7 @@ func TestContainerPrefixes(t *testing.T) {
 				store.ContainerInfo{ID: "cid2", Name: cNamePrefix2},
 			},
 		}))
-	state.UpsertManifestTarget(testutils.NewManifestTargetWithPod(
+	state.UpsertManifestTarget(manifestutils.NewManifestTargetWithPod(
 		model.Manifest{Name: "singleContainer"},
 		// Pod with just one container -- logs should NOT be prefixed with container name
 		store.Pod{
@@ -220,7 +220,7 @@ func TestLogsByPodPhase(t *testing.T) {
 
 			state := f.store.LockMutableStateForTesting()
 			state.WatchFiles = true
-			state.UpsertManifestTarget(testutils.NewManifestTargetWithPod(
+			state.UpsertManifestTarget(manifestutils.NewManifestTargetWithPod(
 				model.Manifest{Name: "server"},
 				store.Pod{
 					PodID:         podID,
